+++ packages/bpk-component-modal/src/BpkModalScrim.js
@@ -3,14 +3,14 @@ import { TransitionInitialMount } from 'bpk-react-utils';
 
 import './bpk-modal.scss';
 
-const BpkModalScrim = () => {
-  const className = 'bpk-modal-scrim';
-
-  return (
-    <TransitionInitialMount classNamePrefix={className} transitionTimeout={200}>
-      <div className={className} />
-    </TransitionInitialMount>
-  );
-};
+const BpkModalScrim = () => (
+  <TransitionInitialMount
+    appearClassName="bpk-modal-scrim--appear"
+    appearActiveClassName="bpk-modal-scrim--appear-active"
+    transitionTimeout={200}
+  >
+    <div className="bpk-modal-scrim" />
+  </TransitionInitialMount>
+);
 
 export default BpkModalScrim;